Add Interface Association Descriptors.
This was necessary but not sufficient to fix Windows issues with the CDC
serial device, but it may be a good idea anyway from a clarity
standpoint to explicitly group interfaces into functions.

Note that I ran into an issue with chopstx where the device would hang
retrieving the config descriptor when its length was an even multiple of
64 (ENDP0's max packet size).  Adding an IAD for the MIDI interfaces as
well got it out of that edge case, and again, probably makes sense from
a clarity standpoint.