Add a Fritzing file for this project's hardware.
I had found a Blue Pill Fritzing part online a while back, but it was
lacking schematic and pcb svg's.  Today I looked again and found one
that had schematic and pcb views, but the breadboard view had an issue
where the VBat pin didn't want to connect.  I exported the svg's for the
schematic and pcb views and added them to the part I had previously
found, and used this new "merged" part to replace the old one in a
breadboard mock-up I made in Fritzing.

Note that the PCB and schematic views are completely unrouted, and that
I still need to find or make a part for a DA-15S connector (instead of
the current "mystery part").