Check the CCER enable bit in interrupt routine.
It seems that disabling the capture/compare through the CCER register
does not actually disable the output compare functionality.  Even
disabling the interrupt in DIER only disables the firing of the
interrupt, not the setting of the flag in the SR register.  So, only
treat the interrupt as having fired if both the CCER and SR flags are
set.  Should probably also clear the SR flag when setting the CCER flag
to avoid any old triggers from hanging around.