Macro-ize button handling, and comment bitwise-NOT.
Started out just wanting to comment on why the GPIO was complemented
(bitwise-NOT) (because buttons are pulled-UP, thus 1 when not pressed).
Noticed that the button handling was highly repetitive, so turned it
into a macro that is invoked four times right in a row, once for each of
the four buttons.