Always set 'bytes' at the end of the 'case'.
Not only is it more consistent, but seems to cause GCC to generate
smaller code.

Also, it appears that the brackets around the 'if' with the 'case'
inside it are unnecessary, as a 'case' is a label, and labels are part
of a statement, not a statement on their own.